Use npm packages and vite as bundler.
