--- js/leaflet-control-geocoder-1.13.0/.eslintignore
@@ -1,4 +1,0 @@
-bundle.js
-dist/
-node_modules/
-spec/vendor/openlocationcode.js
