--- js/leaflet-control-geocoder-1.13.0/.eslintrc.yaml
@@ -1,15 +1,0 @@
-extends:
-  - eslint:recommended
-
-plugins:
-  - prettier
-
-env:
-  browser: true
-  es6: true
-
-parserOptions:
-  sourceType: module
-
-rules:
-  prettier/prettier: warn
