--- js/leaflet-control-geocoder-1.13.0/.gitignore
@@ -1,5 +1,0 @@
-bundle.js
-demo-rollup/package-lock.json
-demo-webpack/package-lock.json
-dist/
-node_modules/
