--- js/leaflet-control-geocoder-1.13.0/.prettierrc.yaml
@@ -1,2 +1,0 @@
-printWidth: 100
-singleQuote: true
