--- js/leaflet-locatecontrol-0.66.1/L.Control.Locate.css
@@ -1,33 +1,0 @@
-/* Compatible with Leaflet 0.7 */
-.leaflet-control-locate a {
-  font-size: 1.4em;
-  color: #444;
-  cursor: pointer;
-}
-.leaflet-control-locate.active a {
-  color: #2074B6;
-}
-.leaflet-control-locate.active.following a {
-  color: #FC8428;
-}
-
-.leafet-control-locate-location circle {
-  animation: leaflet-control-locate-throb 4s ease infinite;
-}
-
-@keyframes leaflet-control-locate-throb {
-  0% {
-    r: 9;
-    stroke-width: 1;
-  }
-  50% {
-    r: 7;
-    stroke-width: 3;
-  }
-  100% {
-    r: 9;
-    stroke-width: 1;
-  }
-}
-
-/*# sourceMappingURL=L.Control.Locate.css.map */
