--- js/leaflet-polylineDecorator-1.6.0/.gitignore
@@ -1,5 +1,0 @@
-*.sh
-*.ps1
-*~
-*.sublime-*
-node_modules
